requirements.txt
# ── Core agent ────────────────────────────────────────────────────────────────
anthropic>=0.40.0
rich>=13.0.0

# ── Memory ────────────────────────────────────────────────────────────────────
chromadb>=0.5.0
python-frontmatter>=1.0.0

# ── Voice ─────────────────────────────────────────────────────────────────────
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0
pynput>=1.7.6
faster-whisper>=1.0.0
edge-tts>=6.1.9

# ── Browser extension bridge ──────────────────────────────────────────────────
websockets>=12.0
aiohttp>=3.9.0
requests>=2.31.0

# ── Servers (approval + file server) ─────────────────────────────────────────
flask>=3.0.0
werkzeug>=3.0.0
